create default preset and move necessary files
