+++ packages/gluestick-preset-default/.babelrc
@@ -1,0 +1,13 @@
+{
+  "presets": [
+    ["env", {
+      "targets": {
+        "node": 6.3
+      }
+    }],
+    "stage-0"
+  ],
+  "plugins": [
+    "transform-flow-strip-types"
+  ]
+}
