+++ packages/gluestick-preset-default/.npmignore
@@ -1,0 +1,5 @@
+**/__mocks__/**
+**/__tests__/**
+src/**
+coverage/**
+.babelrc
