+++ packages/gluestick-preset-default/package.json
@@ -1,0 +1,92 @@
+{
+  "name": "gluestick-preset-default",
+  "version": "1.9.1",
+  "description": "The default preset for GlueStick project",
+  "main": "build/index.js",
+  "homepage": "https://github.com/TrueCar/gluestick",
+  "bugs": "https://github.com/TrueCar/gluestick/issues",
+  "license": "MIT",
+  "scripts": {
+    "test": "jest",
+    "test-coverage": "jest --coverage"
+  },
+  "keywords": [
+    "react",
+    "redux",
+    "jest",
+    "es6",
+    "es2015",
+    "universal",
+    "isomorphic",
+    "generator",
+    "boilerplate"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:TrueCar/gluestick.git"
+  },
+  "author": "Todd Williams <toddsurfs@icloud.com>",
+  "dependencies": {
+    "mkdirp": "^0.5.1"
+  },
+  "devDependencies": {
+    "babel-plugin-transform-decorators-legacy": "1.3.4",
+    "babel-plugin-transform-flow-strip-types": "^6.22.0",
+    "babel-preset-es2015": "6.22.0",
+    "babel-preset-react": "6.22.0",
+    "babel-preset-stage-0": "6.22.0",
+    "jest": "19.0.0"
+  },
+  "jest": {
+    "notify": true,
+    "testRegex": ".*/__tests__/.*\\.test\\.js$",
+    "roots": [
+      "src"
+    ]
+  },
+  "gsProjectDependencies": {
+    "dependencies": {
+      "babel-loader": "7.0.0",
+      "babel-plugin-transform-decorators-legacy": "1.3.4",
+      "babel-plugin-transform-flow-strip-types": "6.22.0",
+      "babel-polyfill": "6.23.0",
+      "babel-preset-es2015": "6.24.1",
+      "babel-preset-react": "6.24.1",
+      "babel-preset-stage-0": "6.24.1",
+      "css-loader": "0.28.1",
+      "extract-text-webpack-plugin": "2.1.0",
+      "file-loader": "0.11.1",
+      "image-webpack-loader": "3.3.1",
+      "node-sass": "4.5.1",
+      "normalize.css": "^5.0.0",
+      "optimize-css-assets-webpack-plugin": "1.3.1",
+      "postcss-calc": "5.3.1",
+      "postcss-custom-properties": "5.0.2",
+      "postcss-loader": "1.3.3",
+      "react": "15.4.2",
+      "react-dom": "15.4.2",
+      "react-helmet": "4.0.0",
+      "react-hot-loader": "^3.0.0-beta.6",
+      "react-redux": "5.0.4",
+      "react-router": "3.0.2",
+      "react-router-scroll": "0.4.2",
+      "redux": "3.6.0",
+      "redux-thunk": "2.2.0",
+      "sass-loader": "6.0.3",
+      "style-loader": "0.16.1",
+      "url-loader": "0.5.8"
+    },
+    "devDependencies": {
+      "babel-eslint": "7.1.1",
+      "babel-jest": "18.0.0",
+      "babel-plugin-react-transform": "2.0.2",
+      "enzyme": "2.7.1",
+      "eslint": "3.14.1",
+      "eslint-plugin-react": "6.9.0",
+      "flow-bin": "0.45.0",
+      "flow-typed": "^2.0.0",
+      "react-addons-test-utils": "15.4.2",
+      "redbox-react": "1.3.3"
+    }
+  }
+}
