+++ packages/gluestick-preset-default/src/templates/Home.js
@@ -1,7 +1,7 @@
 /* @flow */
-import type { CreateTemplate } from '../../types';
+import type { CreateTemplate } from '../types';
 
-module.exports = (createTemplate: CreateTemplate) => createTemplate`
+module.exports.source = (createTemplate: CreateTemplate) => createTemplate`
 /* @flow */
 
 import React, { Component } from "react";
@@ -40,4 +40,21 @@ export default class Home extends Component {
     );
   }
 }
+`;
+
+module.exports.test = (createTemplate: CreateTemplate) => createTemplate`
+/* @flow */
+
+import React from "react";
+import { shallow } from "enzyme";
+
+import Home from "../Home";
+
+describe("components/Home", () => {
+  it("renders without an issue", () => {
+    const subject = <Home />;
+    const wrapper = shallow(subject);
+    expect(wrapper).toBeDefined();
+  });
+});
 `;
