+++ packages/gluestick-preset-default/src/templates/HomeApp.js
@@ -1,7 +1,7 @@
 /* @flow */
-import type { CreateTemplate } from '../../types';
+import type { CreateTemplate } from '../types';
 
-module.exports = (createTemplate: CreateTemplate) => createTemplate`
+module.exports.source = (createTemplate: CreateTemplate) => createTemplate`
 /* @flow */
 
 import React, { Component } from "react";
@@ -43,4 +43,21 @@ export default connect(
   (dispatch) => bindActionCreators({/** _INSERT_ACTION_CREATORS_ **/}, dispatch)
 )(HomeApp);
 
+`;
+
+module.exports.test = (createTemplate: CreateTemplate) => createTemplate`
+/* @flow */
+
+import React from "react";
+import { shallow } from "enzyme";
+
+import { HomeApp } from "../HomeApp";
+
+describe("containers/HomeApp", () => {
+  it("renders without an issue", () => {
+    const subject = <HomeApp />;
+    const wrapper = shallow(subject);
+    expect(wrapper).toBeDefined();
+  });
+});
 `;
