+++ packages/gluestick-preset-default/src/templates/HomeCss.js
@@ -1,5 +1,5 @@
 /* @flow */
-import type { CreateTemplate } from '../../types';
+import type { CreateTemplate } from '../types';
 
 module.exports = (createTemplate: CreateTemplate) => createTemplate`
 .Home-header {
