+++ packages/gluestick-preset-default/src/templates/babelrc.js
@@ -1,0 +1,12 @@
+/* @flow */
+import type { CreateTemplate } from '../types';
+
+module.exports = (createTemplate: CreateTemplate) => createTemplate`
+{
+  "presets": [
+    "react",
+    "es2015",
+    "stage-0"
+  ]
+}
+`;
