+++ packages/gluestick-preset-default/src/templates/empty.js
