--- packages/gluestick/src/generator/templates/MasterLayoutTest.js
@@ -1,19 +1,0 @@
-/* @flow */
-import type { CreateTemplate } from '../../types';
-
-module.exports = (createTemplate: CreateTemplate) => createTemplate`
-/* @flow */
-
-import React from "react";
-import { shallow } from "enzyme";
-
-import MasterLayout from "../MasterLayout";
-
-describe("components/MasterLayout", () => {
-  it("renders without an issue", () => {
-    const subject = <MasterLayout />;
-    const wrapper = shallow(subject);
-    expect(wrapper).toBeDefined();
-  });
-});
-`;
