--- packages/gluestick/src/generator/templates/babelrc.js
@@ -1,28 +1,0 @@
-/* @flow */
-import type { CreateTemplate } from '../../types';
-
-module.exports = (createTemplate: CreateTemplate) => createTemplate`
-{
-  "presets": [
-    "react",
-    "es2015",
-    "stage-0"
-  ],
-  "env": {
-    "development": {
-      "plugins": [
-        ["react-transform", {
-          "transforms": [{
-            "transform": "react-transform-hmr",
-            "imports": ["react"],
-            "locals": ["module"]
-          }, {
-            "transform": "react-transform-catch-errors",
-            "imports": ["react", "redbox-react"]
-          }]
-        }]
-      ]
-    }
-  }
-}
-`;
