+++ packages/gluestick/src/generators/predefined/clientEntryInit.js
