+++ packages/gluestick/src/generators/predefined/serverEntries.js
