+++ packages/gluestick/src/generators/templates/500hbs.js
