+++ packages/gluestick/src/generators/templates/Dockerfile.js
