+++ packages/gluestick/src/generators/templates/EntryWrapper.js
