+++ packages/gluestick/src/generators/templates/application.js
