+++ packages/gluestick/src/generators/templates/application.server.js
