+++ packages/gluestick/src/generators/templates/dockerignore.js
