+++ packages/gluestick/src/generators/templates/gitignore.js
