+++ packages/gluestick/src/generators/templates/gluestick.config.js
