+++ packages/gluestick/src/generators/templates/gluestick.hooks.js
