+++ packages/gluestick/src/generators/templates/gluestick.plugins.js
