+++ packages/gluestick/src/generators/templates/init.browser.js
