+++ packages/gluestick/src/generators/templates/redux-middleware.js
