+++ packages/gluestick/src/generators/templates/webpack-additions.js
