+++ packages/gluestick/src/generators/templates/webpack.hooks.js
